Fix description of create_sub_devices()
The parameter to `create_sub_devices()` for `info::partition_property::partition_equally` is the number of compute units for each sub device (not the number of sub devices to create).  This matches the existing DPC++ implementation.  It also matches all other existing SYCL implementations, matches the OpenCL behavior, and the SYCL 2020 spec has been updated to make this clear.

I also clarified that that elements of the `counts` array for `info::partition_property::partition_by_counts` are the number of compute units for each created sub device.